fix: TypeError Decimal+=float - all money fields now Numeric(12,2)

+++ schemas/wallet.py
@@ -1,13 +1,17 @@
 from pydantic import BaseModel
 from datetime import datetime
+from decimal import Decimal
 from typing import Optional
 
+
 class AddMoneyRequest(BaseModel):
-    amount: float
+    # Decimal keeps exact precision — no float rounding errors for money
+    amount: Decimal
+
 
 class PayUInitResponse(BaseModel):
     txnid: str
-    amount: float
+    amount: Decimal
     productinfo: str
     firstname: str
     email: str
@@ -18,20 +22,23 @@ class PayUInitResponse(BaseModel):
     key: str
     action: str
 
+
 class WithdrawalRequest(BaseModel):
-    amount: float
+    amount: Decimal
     upi_id: str
+
 
 class WalletTransactionResponse(BaseModel):
     id: int
-    amount: float
+    amount: Decimal
     transaction_type: str
     status: str
     reference_id: Optional[str] = None
     created_at: datetime
-    
+
     class Config:
         from_attributes = True
 
+
 class WalletBalanceResponse(BaseModel):
-    balance: float
+    balance: Decimal
